+++ src/utils/utils.js
@@ -4,6 +4,7 @@ import { BigNumber } from 'bignumber.js'
 import logdown from 'logdown'
 import Web3 from 'web3'
 import moment from 'moment'
+import { fetchFile } from './fetchFile'
 import { isObservableArray } from 'mobx'
 
 const logger = logdown('TW:utils:utils')
@@ -45,22 +46,6 @@ export const getNetworkID = () => {
 
 export function setFlatFileContentToState(file) {
   return fetchFile(file)
-}
-
-export function fetchFile(path) {
-  return new Promise((resolve, reject) => {
-    const rawFile = new XMLHttpRequest()
-
-    rawFile.addEventListener('error', reject)
-    rawFile.open('GET', path, true)
-    rawFile.onreadystatechange = function() {
-      if (rawFile.readyState === 4 && (rawFile.status === 200 || rawFile.status === 0)) {
-        let allText = rawFile.responseText
-        resolve(allText)
-      }
-    }
-    rawFile.send(null)
-  })
 }
 
 export const dateToTimestamp = date => new Date(date).getTime()
