+++ test/stores/Utils.spec.js
@@ -8,6 +8,7 @@ describe(`Utils in the store`, () => {
     // Given
     await getCrowdsaleAssets(3)
     const contractStore = new ContractStore()
+
     // When
     const { DutchProxy, MintedCappedProxy } = contractStore
